requirements.txt
flask
supabase
langchain
langchain-openai
python-dotenv
gunicorn